reset marking system on the first question
Marking system local.storage is reset on the first question and the last question. For the first question this is done using a new boolean variable on the Question type called "firstQuestion". The last question does not have a nextQuestionId assigned to the object, by checking this we knnow it's the final question, the EndButotn algorithm is run instead of the NextQuestion algorithm which rests the mark systems local.storage.